+++ api.js
@@ -7,10 +7,10 @@ const api = axios.create({
   headers: {
     "Content-Type": "application/json",
   },
-  withCredentials: true, // 쿠키 필요 시
+  withCredentials: true, // 서버 쿠키 필요시
 });
 
-// 🔹 요청 인터셉터
+// ✅ 요청 인터셉터: SecureStore에서 accessToken 꺼내기
 api.interceptors.request.use(
   async (config) => {
     const token = await SecureStore.getItemAsync("accessToken");
@@ -22,43 +22,40 @@ api.interceptors.request.use(
   (error) => Promise.reject(error)
 );
 
-// 🔹 응답 인터셉터 (자동 재발급)
+// ✅ 응답 인터셉터: 자동 토큰 재발급
 api.interceptors.response.use(
-  (response) => response, // 정상 응답은 그대로
+  (response) => response,
   async (error) => {
     const originalRequest = error.config;
 
-    // AccessToken 만료라고 판단하는 조건 (예: 401 Unauthorized)
+    // AccessToken 만료일 경우
     if (
       error.response &&
       error.response.status === 401 &&
       !originalRequest._retry
     ) {
       originalRequest._retry = true;
-
       try {
-        // refresh_token API 호출해서 새 Access Token 발급
+        // Refresh API 호출
         const refreshResponse = await axios.post(
           `${API_BASE_URL}/users/refresh_token`,
           {},
           {
-            withCredentials: true, // 서버에서 쿠키에 저장된 refresh_token 사용
+            withCredentials: true, // 서버에서 refresh_token 쿠키 사용
           }
         );
-        console.log("새 Access Token 값:", refreshResponse.data.access_token);
-        console.log("=== refreshResponse.data ===", refreshResponse.data);
-        console.log("=== typeof ===", typeof refreshResponse.data);
+
         const newAccessToken = refreshResponse.data.access_token;
 
-        // 새 Access Token 저장
+        // 새 accessToken 저장
         await SecureStore.setItemAsync("accessToken", newAccessToken);
 
-        // 새 토큰으로 Authorization 헤더 갱신 후 재요청
+        // Authorization 헤더 갱신 후 재요청
         originalRequest.headers.Authorization = `Bearer ${newAccessToken}`;
         return api(originalRequest);
       } catch (refreshError) {
-        // RefreshToken도 만료되었으면 로그아웃 처리
         console.log("Refresh Token 만료:", refreshError);
+        // refresh도 실패하면 reject
         return Promise.reject(refreshError);
       }
     }
